requirements.txt
flask==2.0.2
azure-storage-blob==12.8.1
werkzeug==2.0.2
